feat: Restructure project and add frontend directory
- Create new frontend directory
- Move existing frontend files into the new frontend directory
